+++ components/create-character/CChar-Form-2.vue
@@ -1,50 +1,62 @@
-// Alignment and Faith
+// Appearance
 
 <script setup lang="ts">
 import type { FormError, FormSubmitEvent } from "#ui/types";
 import { useAuthStore } from "#imports";
 
 type CharacterForm = {
-  alignment: string;
-  faith: string;
+  hair: string;
+  skin: string;
+  eyes: string;
+  height: string;
+  weight: number;
+  age: number;
+  gender: string;
 };
 const authStore = useAuthStore();
 const toast = useToast();
 const router = useRouter();
 
-const alignments = [
-  "Chaotic Evil",
-  "Chaotic Good",
-  "Chaotic Neutral",
-  "Lawful Evil",
-  "Lawful Good",
-  "Lawful Neutral",
-  "Neutral",
-  "Neutral Evil",
-  "Neutral Good",
-];
-
 const state = reactive({
-  alignment: undefined,
-  faith: undefined,
+  hair: undefined,
+  skin: undefined,
+  eyes: undefined,
+  age: undefined,
+  height: undefined,
+  weight: undefined,
+  gender: undefined,
 });
 
 const disabled = ref(false);
 const loading = ref(false);
 const buttonText = reactive({
-  nextBtn: "Next: Appearance",
+  nextBtn: "Create Character",
   prevBtn: "Go Back",
 });
 
 const validate = (state: CharacterForm): FormError[] => {
   const errors = [];
-  if (!state.alignment) errors.push({ path: "name", message: "Required" });
+  if (state.hair && state.hair.length > 25)
+    errors.push({ path: "hair", message: "Must be less than 25 characters" });
+  if (state.skin && state.skin.length > 25)
+    errors.push({ path: "skin", message: "Must be less than 25 characters" });
+  if (state.eyes && state.eyes.length > 25)
+    errors.push({ path: "eyes", message: "Must be less than 25 characters" });
+  if (state.height && state.height.length > 25)
+    errors.push({ path: "height", message: "Must be less than 25 characters" });
+  if (state.gender && state.gender.length > 25)
+    errors.push({ path: "gender", message: "Must be less than 25 characters" });
+  if (state.age && state.age < 1)
+    errors.push({ path: "age", message: "Must be at least 1 year old" });
+  if (state.weight && state.weight < 1)
+    errors.push({ path: "weight", message: "Must be at least 1lb" });
   return errors;
 };
 
-const goBack = () => {
+async function goBack() {
+  await authStore.ensureValidToken();
   router.back();
-};
+}
 
 async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
   disabled.value = true;
@@ -62,12 +74,21 @@ async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
         "Content-Type": "application/json",
       },
       body: {
-        alignment: event.data.alignment,
-        faith: event.data.faith,
+        hair: event.data.hair,
+        skin: event.data.skin,
+        eyes: event.data.eyes,
+        age: event.data.age,
+        height: event.data.height,
+        weight: event.data.weight,
+        gender: event.data.gender,
       },
     });
     console.log(response);
-    await navigateTo("/cchar/cchar-3");
+    toast.add({
+      title: "Character Created!",
+      icon: "i-heroicons-check-circle-solid",
+    });
+    await navigateTo("/");
   } catch (err) {
     console.error(err);
     toast.add({
@@ -78,30 +99,51 @@ async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
   } finally {
     loading.value = false;
     disabled.value = false;
-    buttonText.nextBtn = "Next: Appearance";
+    buttonText.nextBtn = "Create Character";
     buttonText.prevBtn = "Back";
-    state.alignment = undefined;
-    state.faith = undefined;
+    state.hair = undefined;
+    state.skin = undefined;
+    state.eyes = undefined;
+    state.age = undefined;
+    state.height = undefined;
+    state.weight = undefined;
+    state.gender = undefined;
   }
 }
 </script>
 
 <template>
   <UForm :validate="validate" :state="state" @submit.prevent="onSubmit">
-    <UFormGroup label="Alignment" name="alignment" class="mb-4">
-      <USelect
-        v-model="state.alignment"
-        :options="alignments"
-        :disabled="disabled"
-        required
-      />
+    <div class="grid grid-cols-2 gap-4">
+      <UFormGroup label="Hair" name="hair">
+        <UInput v-model="state.hair" :disabled="disabled" type="text" />
+      </UFormGroup>
+      <UFormGroup label="Skin" name="skin">
+        <UInput v-model="state.skin" :disabled="disabled" type="text" />
+      </UFormGroup>
+      <UFormGroup label="Eyes" name="eyes">
+        <UInput v-model="state.eyes" :disabled="disabled" type="text" />
+      </UFormGroup>
+      <UFormGroup label="Age (years)" name="age">
+        <UInput v-model="state.age" :disabled="disabled" type="number" />
+      </UFormGroup>
+      <UFormGroup label="Height" name="height">
+        <UInput v-model="state.height" :disabled="disabled" type="text" />
+      </UFormGroup>
+      <UFormGroup label="Weight (lbs)" name="weight" class="mb-4">
+        <UInput v-model="state.weight" :disabled="disabled" type="number" />
+      </UFormGroup>
+    </div>
+    <UFormGroup label="Gender" name="gender">
+      <UInput v-model="state.gender" :disabled="disabled" type="text" />
     </UFormGroup>
-    <UFormGroup label="Faith" name="faith">
-      <UInput v-model="state.faith" :disabled="disabled" type="text" />
-    </UFormGroup>
-    <div class="flex gap-8 justify-center">
+
+    <div class="flex flex-col justify-center">
       <UButton
-        class="p-2 box-border w-2/5 text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
+        block
+        color="amber"
+        variant="solid"
+        class="p-2 box-border text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
         :loading="loading"
         @click="goBack"
       >
@@ -109,7 +151,8 @@ async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
       >
       <UButton
         type="submit"
-        class="p-2 box-border w-2/5 text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
+        block
+        class="p-2 box-border text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
         :loading="loading"
       >
         {{ buttonText.nextBtn }}</UButton
